+++ CHANGELOG.md
@@ -4,6 +4,18 @@ Notable changes to Conduit are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project aspires to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+#### Conduit
+
+#### Blueprint
+- Fixed partitioner so it reverses vertex order as needed in polyhedral face definitions when extracting mesh elements.
+
+#### Relay
+
+
 ## [0.9.1] - Released 2024-02-09
 
 ### Changed